fix: vite HRM crashing issue due to context
