+++ client/src/context.js
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+
+export const TransactionContext = createContext()
+export const SwingContext = createContext()
