+++ client/src/hooks/use-web3.js
@@ -1,6 +1,6 @@
 import { useContext } from 'react'
 
-import { TransactionContext } from '../context/TransactionContext'
+import { TransactionContext } from '../context'
 
 export default () => {
   return useContext(TransactionContext)
