+++ client/src/main.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import './index.css'
 import App from './App'
 
-import { TransactionProvider } from './context/TransactionContext'
+import TransactionProvider from './providers/TransactionProvider'
 
 const rootElement = document.getElementById('root')
